Extract CoreData managed model classes to its own file
